requirements.txt
# Core dependencies
pyautogen>=0.2.0
openai>=1.1.0
python-dotenv>=1.0.0

# Database handling
sqlite-utils

# Utilities
pyyaml>=6.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
rich>=13.0.0

# Evaluation
pytest>=7.0.0